adds serverless setup file, #10
