+++ lambda/templates/html-template.hbs
@@ -12,7 +12,7 @@
     </tr>
     {{#each Items}}
       <tr>
-        <td>{{objectName.S}}</td>
+        <td>{{title.S}}</td>
         <td>{{objectId.S}}</td>
         <td>
           {{#if active.BOOL}}
