+++ lambda/toggleStatus/index.js
@@ -1,0 +1,28 @@
+const AWS = require("aws-sdk");
+const dynamoDB = new AWS.DynamoDB({region: 'eu-west-2'});
+const querystring = require('querystring');
+const lambda = new AWS.Lambda({region: 'eu-west-2'});
+const listObjects = require('../lib/listObjects');
+
+module.exports.handler = function(event, context, callback) {
+  var params = querystring.parse(event.body);
+
+  return dynamoDB.updateItem({
+    TableName: process.env.TABLE_NAME,
+    UpdateExpression: "SET active = :a",
+    ExpressionAttributeValues: {
+     ":a": {
+       BOOL: params.active === "true" ? true : false
+      }
+    },
+    Key: {
+     "id": {
+       S: params.id
+      }
+    }
+  }, function(err, data) {
+    if (err) return callback(err);
+
+    return listObjects({}, {}, callback);
+  });
+}
