+++ lambda/toggleStatus/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "togglestatus",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
